fix error for modify password + delete excess file + fix link redirection
