+++ php/class/user.php
@@ -29,25 +29,29 @@ class User
     public $firstname;
     public $lastname;
     public $password;
+    public $role;
 
-    public function __construct($id, $email, $firstname, $lastname, $password,)
+    public function __construct($id, $email, $firstname, $lastname, $password, $role)
     {
         $this->id = $id;
         $this->email = $email;
         $this->firstname = $firstname;
         $this->lastname = $lastname;
         $this->password = $password;
+        $this->role = $role;
     }
 
     public function getAllinfo()
     {
-        return $this->password;
+        return $this->id;
         return $this->email;
         return $this->firstname;
         return $this->lastname;
+        return $this->password;
+        return $this->role;
     }
 
-    public function getId($id)
+    public function getId()
     {
         return $this->id;
     }
@@ -56,17 +60,7 @@ class User
         $this->id = $id;
     }
 
-    public function getPassword($password)
-    {
-        return $this->password;
-    }
-    public function setPassword($password)
-    {
-        $this->password = $password;
-    }
-
-
-    public function getEmail($email)
+    public function getEmail()
     {
         return $this->email;
     }
@@ -75,8 +69,7 @@ class User
         $this->email = $email;
     }
 
-
-    public function getFirstname($firstname)
+    public function getFirstname()
     {
         return $this->firstname;
     }
@@ -85,8 +78,7 @@ class User
         $this->firstname = $firstname;
     }
 
-
-    public function getLastname($lastname)
+    public function getLastname()
     {
         return $this->lastname;
     }
@@ -95,10 +87,27 @@ class User
         $this->lastname = $lastname;
     }
 
+    public function getPassword()
+    {
+        return $this->password;
+    }
+    public function setPassword($password)
+    {
+        $this->password = $password;
+    }
+
+    public function getRole()
+    {
+        return $this->role;
+    }
+    public function setRole($role)
+    {
+        $this->role = $role;
+    }
+
     public function register($bdd)
     {
-
-        $recupUser = $bdd->prepare("SELECT * FROM users WHERE email = ?");
+        $recupUser = $bdd->prepare("SELECT email FROM users WHERE email = ?");
         $recupUser->execute([$this->email]);
         $insertUser = $bdd->prepare("INSERT INTO users (email,lastname,firstname,password) VALUES(?,?,?,?)");
 
@@ -109,10 +118,11 @@ class User
         } elseif (lastname($this->lastname) == false) {
         } elseif (same_password($this->password, $_POST['confirm_password']) == false) {
         } elseif ($recupUser->rowCount() > 0) {
-            echo 'Email déjà utilisé';
+            $_SESSION['message'] = 'Email déjà utilisé';
         } else {
+            unset($_SESSION['message']);
             $insertUser->execute([$this->email, $this->lastname, $this->firstname, password_hash($this->password, PASSWORD_DEFAULT)]);
-            header('Location:../index.php');
+            header('Location: ./connect.php');
         }
     }
 
@@ -121,17 +131,16 @@ class User
     {
         $request = $bdd->prepare("SELECT * FROM users WHERE email = ?");
         $request->execute([$this->email]);
-        $res = $request->fetchAll(PDO::FETCH_OBJ);
-        // var_dump($res);
-
-        $recupUser = $bdd->prepare("SELECT * FROM users WHERE email = ? AND password = ?");
-        $recupUser->execute([$this->email, $res[0]->password]);
-        $result = $recupUser->fetch(PDO::FETCH_OBJ);
-        // $_SESSION['user'] = serialize($this);
+        $res = $request->fetch(PDO::FETCH_OBJ);
 
         if (email($this->email) == false) {
         } elseif (password($this->password) == false) {
-        } elseif ($recupUser->rowCount() > 0) {
+        } elseif ($request->rowCount() > 0) {
+
+            $recupUser = $bdd->prepare("SELECT * FROM users WHERE email = ? AND password = ?");
+            $recupUser->execute([$this->email, $res->password]);
+            $result = $recupUser->fetch(PDO::FETCH_OBJ);
+
             if ($result != false) {
                 if (password_verify($this->password, $result->password)) {
                     $this->id = $result->id;
@@ -139,13 +148,17 @@ class User
                     $this->firstname = $result->firstname;
                     $this->lastname = $result->lastname;
                     $this->password = $result->password;
+                    $this->role = $result->role;
 
                     $_SESSION['user'] = $this;
+                    unset($_SESSION['message']);
                     header('Location: ../index.php');
                 }
             }
         } else {
-            echo 'uilisateurs inconnu';
+            $_SESSION['message'] = 'utilisateurs inconnu';
+            //? VOIR SI ON PEUT S'EN PASSER
+            header('Location: connect.php');
         }
     }
 
@@ -154,12 +167,10 @@ class User
 
         $request = $bdd->prepare("SELECT * FROM users WHERE id = ?");
         $request->execute([$this->id]);
-        $res = $request->fetchAll(PDO::FETCH_OBJ);
-        // var_dump($res[0]->password);
+        $res = $request->fetch(PDO::FETCH_OBJ);
 
         $recupUser = $bdd->prepare("SELECT * FROM users WHERE email = ? AND id != ?");
         $recupUser->execute([$this->email, $this->id]);
-        // var_dump($res);
         $insertUser = $bdd->prepare("UPDATE users SET email = ?, firstname = ?, lastname = ?, password = ? WHERE id = ? ");
 
         if (email($this->email) == false) {
@@ -167,12 +178,17 @@ class User
         } elseif (firstname($this->firstname) == false) {
         } elseif (lastname($this->lastname) == false) {
         } elseif ($recupUser->rowCount() > 0) {
-            echo 'Email déjà utilisé';
+            $_SESSION['message'] = 'Email déjà utilisé';
+            //? VOIR SI ON PEUT S'EN PASSER
+            header('Location:./profil.php');
         } else {
-            if ($this->password != password_verify($this->password, $res[0]->password)) {
-                echo  "Ce n'est pas le bon mot de passe";
+            if ($this->password != password_verify($this->password, $res->password)) {
+                $_SESSION['message'] = 'Ce n\'est pas le bon mot de passe';
+                //? VOIR SI ON PEUT S'EN PASSER
+                header('Location:./profil.php');
             } else {
-                $insertUser->execute([$this->email, $this->firstname, $this->lastname, $res[0]->password, $this->id]);
+                unset($_SESSION['message']);
+                $insertUser->execute([$this->email, $this->firstname, $this->lastname, $res->password, $this->id]);
                 $_SESSION['user']->email = $this->email;
                 $_SESSION['user']->firstname = $this->firstname;
                 $_SESSION['user']->lastname = $this->lastname;
@@ -181,21 +197,21 @@ class User
             }
         }
     }
+
     public function updatePassword($bdd)
     {
         $request = $bdd->prepare("SELECT * FROM users WHERE id = ?");
         $request->execute([$this->id]);
-        $res = $request->fetchAll(PDO::FETCH_OBJ);
-        // var_dump($res);
+        $res = $request->fetch(PDO::FETCH_OBJ);
         $insertUser = $bdd->prepare("UPDATE users SET password = ? WHERE id = ? ");
 
-
-        if (password($this->password) == false) {
-        } elseif (empty($_POST['new_password'])) {
-            echo 'Champ New Password vide';
-        } elseif ($_POST['password'] != password_verify($_POST['password'], $res[0]->password)) {
-            echo  "Ce n'est pas le bon mot de passe";
+        if (password($_POST['password']) == false) {
+        } elseif (password($this->password) == false) {
+        } elseif ($_POST['password'] != password_verify($_POST['password'], $res->password)) {
+            $_SESSION['message'] = 'Ce n\'est pas le bon mot de passe';
+            header('Location:./modifyPassword.php');
         } else {
+            unset($_SESSION['message']);
             $insertUser->execute([$this->password, $this->id]);
             $_SESSION['user']->password = $this->password;
             header('Location:./profil.php');
@@ -209,32 +225,11 @@ class User
     public function isConnected()
     {
         if (isset($_SESSION['user']->login)) {
-            echo 'Connected';
+            // echo 'Connected';
             return true;
         } else {
-            echo 'Disconnected';
+            // echo 'Disconnected';
             return false;
         }
     }
 }
-
-// $user = new User($login, $password, $email, $firstname, $lastname);
-// $user = new User('a', 'a', 'a', 'a', 'a');
-// var_dump($user);
-
-// $user->register($bdd);
-// $user->connect($bdd);
-// $user->disconnect();
-// $user->isConnected();
-
-// $user->getAllinfo();
-// $user->getLogin();
-// $user->setLogin();
-// $user->getPassword();
-// $user->setPassword();
-// $user->getEmail();
-// $user->setEmail();
-// $user->getFirstname();
-// $user->setFirstname();
-// $user->getLastname();
-// $user->setLastname();
