+++ php/disconnect.php
@@ -2,6 +2,6 @@
 // var_dump('ddd');
 require_once('./class/user.php');
 
-$user = new User('', '', '', '', '');
+$user = new User('', '', '', '', '', '');
 $user->disconnect();
 header('Location: ../index.php');
